+++ CHANGELOG.md
@@ -2,8 +2,13 @@
 
 All notable changes to the "swift-add-documentation" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+## 0.0.2
 
-## [Unreleased]
+- Fixed default hotkeys. They were in the wrong place in the `package.json` file.
+- Added functionality to insert the template after typing `///`.
+- Added functionality to start a new comment line  when using typing `enter+return`.
+- Added functionality to insert a new comment line when using typing `enter+opt+return` that can be used in the middle of the current line and does not break up the existing line.
+
+## 0.0.1
 
 - Initial release